fix race condition when unmounting a volume takes a very long time
Whilst trying to fix this bug

 https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=194296

I discovered that there's a race condition when unmounting a volume
that has lots of outstanding writes. Under Linux, at least, this
results in umount(8) taking very long to complete. However, the
kernel tells us, just after the umount syscall I believe, that
/proc/mounts has changed. So we check that file to see if it was
us, HAL, that unmounted the volume. Because if it wasn't, we want
to clean up the mount point if we created it.

To do this, we first lock /media/.hal-mtab to ensure noone else is
modifying it. But this will block until Unmount() terminates (because
Unmount() is holding the lock on that file) and that can take many
many seconds, e.g. 30 secs for example, if you got a bunch of
outstanding writes.

The fix here is to check whether we're already executing Unmount()
on the device in question. If we are, then, hey, we can infer
that HAL is unmounting that volume, not someone else.
(cherry picked from e35d886bf811baa0c15e9d71a6296f37ef403dde commit)