fix contains_not fdi-directive
This patch fixes the contains_not FDI file directive/rule to return true (let
the match not fail) if the property is not set on the device. This allow to
have this rule:

       <match key="info.addons" contains_not="hald-addon-input">
         <append key="info.addons" type="strlist">hald-addon-input</append>
       </match>

Without this patch this fails if info.addons is not already set. With the
change it would match if:
* info.addons is not available
or
* info.addons contains no string 'hald-addon-input'