fix return type of SetPowerSave()
The spec say about SetPowerSave() on the
org.freedesktop.Hal.Device.SystemPowerManagement interface the return
type should be integer, but in
fdi/policy/10osvendor/10-power-mgmt-policy.fdi the return value type
get specified as boolean.

This patch change the type to Integer as the spec say and the code
allow since it forward the retun value of the called tool to set the
mode.