the runner doesn't handle D-Bus disconnect properly
It appears that if the connection to D-Bus is lost or some such is
lost, HAL doesn't check the result of DBusPendingCall from
dbus_connection_send_with_reply() [1] to see if the result is not NULL
and as such an invalid case is being hit within the D-Bus API and
we're hitting a D-Bus assert and causing a crash.

Provided is a patch to see if it's NULL and handle it properly.